refactor(telegram): extract outbound text delivery into TelegramTextDeliveryMixin (adapter god-file slice)

# Conflicts:
#	plugins/platforms/telegram/adapter.py
